project-3/requirements.txt
tensorflow>=2.12.0
streamlit>=1.32.0
numpy>=1.23.0
Pillow>=9.0.0
opencv-python-headless>=4.7.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
seaborn>=0.12.0
kaggle>=1.5.0
